LayoutTests/http/tests/pasteboard/copy-image-from-context-menu-expected.txt

This tests that using the 'Copy Image' context menu item codepath does not copy the image URL.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS event.clipboardData.items.length is 2
PASS event.clipboardData.items[0].type is 'text/html'
PASS event.clipboardData.items[1].type is 'image/png'
PASS successfullyParsed is true

TEST COMPLETE
To run this test manually, click "Copy Image" in the context menu with the image below, and paste in the box with the black border.

NOTE: This test needs to run on a HTTP server as this bug reproduces only with hosted images.
